requirements-dev.txt
coverage==3.7.1
django-nose==1.4.3
flake8==2.4.1
mccabe==0.2.1
mock==1.0.1
mockldap==0.2.0  # Required for mock_ldap_helpers
nose==1.3.6
pep8==1.5.7
pyflakes==0.8.1
ipdb
